Bruk ny modal fra aksel
